removed alchemy dependency
