+++ frontend/src/views/CreatePollView.vue
@@ -5,7 +5,7 @@ import AppButton from '@/components/AppButton.vue';
 import RemoveIcon from '@/components/RemoveIcon.vue';
 import AddIcon from '@/components/AddIcon.vue';
 import SuccessInfo from '@/components/SuccessInfo.vue';
-import { retry, Pinata, encryptJSON, isERCTokenContract, guessStorageSlot } from '@/utils';
+import { retry, Pinata, encryptJSON } from '@/utils';
 import type { PollManager } from '@oasisprotocol/demo-voting-contracts';
 import { getAddress, parseEther, JsonRpcProvider, getBytes, AbiCoder, Contract } from 'ethers';
 
@@ -13,7 +13,7 @@ import { usePollManager, usePollManagerWithSigner } from '../contracts';
 import { useEthereumStore } from '../stores/ethereum';
 import type { AclOptions, AclOptionsAllowAll, AclOptionsAllowList, AclOptionsToken, AclOptionsXchain, Poll } from '../types';
 import { computedAsync } from '../utils';
-import { tokenDetailsFromProvider, xchain_ChainNamesToChainId, xchainRPC } from "../xchain";
+import { guessStorageSlot, isERCTokenContract, tokenDetailsFromProvider, xchain_ChainNamesToChainId, xchainRPC } from "../xchain";
 
 const eth = useEthereumStore();
 const dao = usePollManager();
